Simplify category functions
