+++ src/template.html
@@ -25,14 +25,14 @@
 </section>
 <section class="categories none">
 <div id="btns">
-<button id="Fire" class="activebtn">Fire</button>
-<button id="Water" class="activebtn">Water</button>
-<button id="Bugs" class="activebtn">Bugs</button>
-<button id="Ground" class="activebtn">Ground</button>
-<button id="Ice" class="activebtn">Ice</button>
-<button id="Electric" class="activebtn">Electric</button>
-<button id="Poison" class="activebtn">Poison</button>
-<button id="Normal" class="activebtn">Normal</button>
+<button id="fire" class="activebtn">Fire</button>
+<button id="water" class="activebtn">Water</button>
+<button id="bugs" class="activebtn">Bugs</button>
+<button id="ground" class="activebtn">Ground</button>
+<button id="ice" class="activebtn">Ice</button>
+<button id="electric" class="activebtn">Electric</button>
+<button id="poison" class="activebtn">Poison</button>
+<button id="normal" class="activebtn">Normal</button>
 <button id="Ghost" class="activebtn">Ghost</button>
 <button id="Fighting" class="activebtn">Fighting</button>
 <button id="Steel" class="activebtn">Steel</button>
